Partial Fix for #5098 - Suggested Values dissapear
* QA Data Query Suggested Values being Randomly Removed
* This is only QA.  Still have to fully vet the disappearance

+++ lib/import.php
@@ -1309,10 +1309,12 @@ function xml_to_data_query($hash, &$xml_array, &$hash_cache, &$files, $replace_s
 			/* import into: snmp_query_graph_sv */
 			if (is_array($item_array['sv_graph'])) {
 				/* if the user choose to replace data query suggested values */
-				if ($data_query_graph_id > 0 && $replace_svalues) {
-					db_execute_prepared('DELETE FROM snmp_query_graph_sv
-						WHERE snmp_query_graph_id = ?',
-						array($data_query_graph_id));
+				if ($data_query_graph_id > 0 && $replace_svalues && cacti_sizeof($item_array['sv_graph'])) {
+					if (!$preview_only) {
+						db_execute_prepared('DELETE FROM snmp_query_graph_sv
+							WHERE snmp_query_graph_id = ?',
+							array($data_query_graph_id));
+					}
 				}
 
 				foreach ($item_array['sv_graph'] as $sub_item_hash => $sub_item_array) {
@@ -1363,10 +1365,12 @@ function xml_to_data_query($hash, &$xml_array, &$hash_cache, &$files, $replace_s
 			/* import into: snmp_query_graph_rrd_sv */
 			if (is_array($item_array['sv_data_source'])) {
 				/* if the user choose to replace data query suggested values */
-				if ($data_query_graph_id > 0 && $replace_svalues) {
-					db_execute_prepared('DELETE FROM snmp_query_graph_rrd_sv
-						WHERE snmp_query_graph_id = ?',
-						array($data_query_graph_id));
+				if ($data_query_graph_id > 0 && $replace_svalues && cacti_sizeof($item_array['sv_data_source'])) {
+					if (!$preview_only) {
+						db_execute_prepared('DELETE FROM snmp_query_graph_rrd_sv
+							WHERE snmp_query_graph_id = ?',
+							array($data_query_graph_id));
+					}
 				}
 
 				foreach ($item_array['sv_data_source'] as $sub_item_hash => $sub_item_array) {
